undo changes in the plugin to be executed now in the test
