+++ RealJenkinsRuleInit/src/main/java/org/jvnet/hudson/test/RealJenkinsRuleInit.java
@@ -24,18 +24,10 @@
 
 package org.jvnet.hudson.test;
 
-import hudson.util.PluginServletFilter;
 import hudson.Plugin;
 import java.net.URL;
 import java.net.URLClassLoader;
 import jenkins.model.Jenkins;
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import java.io.IOException;
 
 public class RealJenkinsRuleInit extends Plugin {
 
@@ -44,30 +36,6 @@ public class RealJenkinsRuleInit extends Plugin {
 
     @Override
     public void start() throws Exception {
-        /*
-        According to ServletRequest documentation:
-            If the parameter data was sent in the request body, such as occurs with an HTTP POST request,
-            then reading the body directly via getInputStream or getReader can interfere with the execution of this method.
-        Check RealJenkinsRule#runRemotely, it has
-         conn.setRequestProperty("Content-Type", "text/plain; utf-8");
-
-         to avoid this issue.
-        */
-        PluginServletFilter.addFilter(new Filter() {
-
-          @Override
-          public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
-              String fake = request.getParameter("fake");
-              chain.doFilter(request, response);
-          }
-
-          @Override
-          public void destroy() {}
-
-          @Override
-          public void init(FilterConfig filterConfig) throws ServletException {}
-        });
-
         new URLClassLoader(new URL[] {new URL(System.getProperty("RealJenkinsRule.location"))}, ClassLoader.getSystemClassLoader().getParent()).
                 loadClass("org.jvnet.hudson.test.RealJenkinsRule$Init2").
                 getMethod("run", Object.class).
